Extract shared types between frontend and backend * DTO => data transfer object * less circular dependencies in modules * remove ISerializable interface
