+++ src/Messaging.ts
@@ -1,8 +1,8 @@
 import { BrowserWindow, ipcMain, ipcRenderer, WebContents } from 'electron';
 import path from 'path';
 import { IImportItem } from './clipper/server';
-import { ID } from './entities/ID';
-import { ITag } from './entities/Tag';
+import { ID } from 'src/api/ID';
+import { ITag } from 'src/api/TagDTO';
 import { ViewMethod } from './frontend/stores/UiStore';
 
 /**
