+++ src/api/ID.ts
@@ -1,0 +1,7 @@
+import { v4 as uuid } from 'uuid';
+
+export type ID = string;
+
+export function generateId(): ID {
+  return uuid();
+}
