+++ src/entities/File.ts
@@ -10,31 +10,9 @@ import Path from 'path';
 import ExifIO from 'common/ExifIO';
 import FileStore from 'src/frontend/stores/FileStore';
 import { FileStats } from 'src/frontend/stores/LocationStore';
-import { ID, IResource, ISerializable } from './ID';
+import { ID } from 'src/api/ID';
 import { ClientTag } from './Tag';
-
-export const IMG_EXTENSIONS = [
-  'gif',
-  'png',
-  'apng', // animated png
-  'jpg',
-  'jpeg',
-  'jfif',
-  'webp',
-  'tif',
-  'tiff',
-  'bmp',
-  'svg',
-  'psd', // Photoshop
-  'kra', // Krita
-  // 'xcf', // Gimp
-  'exr', // OpenEXR
-  // 'raw', there are many RAW file extensions :( https://fileinfo.com/filetypes/camera_raw
-  // 'avif',
-  // 'heic', // not supported by Sharp out of the box https://github.com/lovell/sharp/issues/2871
-  // TODO: 'blend', raw, etc.?
-] as const;
-export type IMG_EXTENSIONS_TYPE = typeof IMG_EXTENSIONS[number];
+import { IFile, IMG_EXTENSIONS_TYPE } from 'src/api/FileDTO';
 
 /** Retrieved file meta data information */
 interface IMetaData {
@@ -50,32 +28,12 @@ interface IMetaData {
   dateCreated: Date;
 }
 
-/* A File as it is represented in the Database */
-export interface IFile extends IMetaData, IResource {
-  /** Identifier for a file that persists after renaming/moving (retrieved from fs.Stats.ino) */
-  ino: string;
-  locationId: ID;
-  /** Path relative to Location */
-  relativePath: string;
-  absolutePath: string;
-  tags: ID[];
-  /** When the file was imported into Allusion */
-  dateAdded: Date;
-  /** When the file was modified in Allusion, not related to OS modified date */
-  dateModified: Date;
-  /**
-   * When the file was last indexed in Allusion: concerning the metadata and thumbnail.
-   * If the system's modified date of the file exceeds this date, those properties shoudld be re-initialized
-   **/
-  dateLastIndexed: Date;
-}
-
 /**
  * A File as it is stored in the Client.
  * It is stored in a MobX store, which can observe changed made to it and subsequently
  * update the entity in the backend.
  */
-export class ClientFile implements ISerializable<IFile> {
+export class ClientFile {
   private store: FileStore;
   private saveHandler: IReactionDisposer;
   private autoSave: boolean = true;
