+++ src/frontend/components/Toaster.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { action, makeObservable, observable } from 'mobx';
 import { observer } from 'mobx-react-lite';
 
-import { ID } from 'src/entities/ID';
+import { ID } from 'src/api/ID';
 
 import { Toast } from 'widgets/notifications';
 import { Button } from 'widgets/Button';
