+++ src/frontend/containers/AppToolbar/Menus.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import { observer } from 'mobx-react-lite';
 
-import { OrderDirection } from 'src/backend/DBRepository';
-
 import { IconSet, KeyCombo } from 'widgets';
 import { MenuButton, MenuRadioGroup, MenuRadioItem } from 'widgets/menus';
 import { getThumbnailSize } from '../ContentView/utils';
 import { MenuDivider, MenuSliderItem } from 'widgets/menus/menu-items';
 import { useStore } from 'src/frontend/contexts/StoreContext';
-import { FileOrder } from 'src/frontend/stores/FileStore';
+import { FileOrder, OrderDirection } from 'src/api/FileDTO';
 
 // Tooltip info
 const enum Tooltip {
