+++ src/frontend/image/ImageLoader.ts
@@ -2,7 +2,8 @@ import fse from 'fs-extra';
 import { action } from 'mobx';
 import ExifIO from 'common/ExifIO';
 import { thumbnailMaxSize } from 'common/config';
-import { ClientFile, IFile, IMG_EXTENSIONS_TYPE } from 'src/entities/File';
+import { ClientFile } from 'src/entities/File';
+import { IFile, IMG_EXTENSIONS_TYPE } from 'src/api/FileDTO';
 import TifLoader from './TifLoader';
 import { generateThumbnailUsingWorker } from './ThumbnailGeneration';
 import StreamZip from 'node-stream-zip';
